django_backend/requirements/requirements.txt
asgiref==3.2.10
bcrypt==3.2.0
boto3==1.16.4
botocore==1.19.4
cffi==1.14.3
cryptography==3.1.1
Django==3.1.2
django-cors-headers==3.5.0
django-extensions==3.0.9
django-spa==0.3.5
django-storages==1.10.1
django-stubs==1.6.0
django-templated-mail==1.1.1
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
jmespath==0.10.0
jsonfield==3.1.0
mypy==0.782
mypy-extensions==0.4.3
paramiko==2.7.2
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
PyNaCl==1.4.0
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
whitenoise==5.0.1
mock==4.0.3

